requirements.txt
gradio==4.44.1
scikit-learn==1.5.2
openai>=1.40.0
huggingface_hub<0.24.0
python-dotenv>=1.0.1
